+++ pkg/api/api.go
@@ -1,0 +1,11 @@
+package api
+
+import "runtime"
+
+func trace() *runtime.Frame {
+	pc := make([]uintptr, 15)
+	n := runtime.Callers(2, pc)
+	frames := runtime.CallersFrames(pc[:n])
+	frame, _ := frames.Next()
+	return &frame
+}
