+++ pkg/validation/report_group_with_members_validation.go
@@ -2,28 +2,15 @@ package validation
 
 import (
 	"database/sql"
-	"errors"
 
 	"excel-report-email-scheduler/pkg/datasource"
 	"excel-report-email-scheduler/pkg/ereserror"
+
+	"github.com/pkg/errors"
 )
 
-type Validation struct {
-	datasource *datasource.MsupplyEresDatasource
-	sqlClient  *datasource.SqlClient
-}
-
-func New(datasource *datasource.MsupplyEresDatasource) (*Validation, error) {
-	sqlClient, err := datasource.NewSqlClient()
-	if err != nil {
-		err = ereserror.New(500, err, "Could not open database")
-		return nil, err
-	}
-
-	return &Validation{datasource: datasource, sqlClient: sqlClient}, nil
-}
-
 func (validator *Validation) ReportGroupDuplicates(reportGroupWithMembers datasource.ReportGroupWithMembersRequest) error {
+	frame := trace()
 	var id string
 	row := validator.sqlClient.Db.QueryRow("SELECT id FROM ReportGroup WHERE name = $1 LIMIT 1", reportGroupWithMembers.Name)
 
@@ -32,7 +19,7 @@ func (validator *Validation) ReportGroupDuplicates(reportGroupWithMembers dataso
 		return nil
 	default:
 		if reportGroupWithMembers.ID == "" || (reportGroupWithMembers.ID != id) {
-			err = ereserror.New(500, err, "Cannot have more than one report groups with same name")
+			err = ereserror.New(500, errors.Wrap(err, frame.Function), "Cannot have more than one report groups with same name")
 			return err
 		}
 	}
@@ -40,10 +27,11 @@ func (validator *Validation) ReportGroupDuplicates(reportGroupWithMembers dataso
 }
 
 func (validator *Validation) ReportGroupMustHaveMembers(reportGroupWithMembers datasource.ReportGroupWithMembersRequest) error {
+	frame := trace()
 	memberLength := len(reportGroupWithMembers.Members)
 	if memberLength <= 0 {
 		err := errors.New("report group must have at least one member")
-		err = ereserror.New(500, err, err.Error())
+		err = ereserror.New(500, errors.Wrap(err, frame.Function), err.Error())
 		return err
 	}
 
@@ -52,10 +40,10 @@ func (validator *Validation) ReportGroupMustHaveMembers(reportGroupWithMembers d
 
 func (validator *Validation) GroupMemberUserIDsMustHaveElements(groupMemberUserIDs []string) error {
 	memberLength := len(groupMemberUserIDs)
-
+	frame := trace()
 	if memberLength <= 0 {
 		err := errors.New("report group must have members")
-		err = ereserror.New(500, err, err.Error())
+		err = ereserror.New(500, errors.Wrap(err, frame.Function), err.Error())
 		return err
 	}
